created orderSummary component and utilized back button component
