+++ src/app/favorites/styles.ts
@@ -1,0 +1,73 @@
+import styled, { createGlobalStyle } from 'styled-components';
+
+import { Heart } from 'react-feather';
+
+import NavBar from '../../components/NavBarFolder/NavBar';
+
+export const GlobalStyle = createGlobalStyle`
+  body {
+    background:white;
+    color: black;
+    overflow: visible;
+    
+  }
+`;
+
+export const FavoriteDiv = styled.div`
+  display: flex;
+  flex-direction: row;
+  align-items: start;
+  justify-content: space-around;
+  width: 100%;
+  margin-bottom: 50px;
+  margin-top: 30px;
+`;
+
+export const OutterFavoriteDiv = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  border-radius: 10px;
+  background: var(--White, #fff);
+  box-shadow: 0px 1px 4px 1px rgba(0, 0, 0, 0.2);
+  width: 800px;
+  height: 700px;
+  overflow: scroll;
+  margin-top: 10px;
+`;
+
+export const BackDiv = styled.button`
+  display: flex;
+  flex-direction: row;
+  align: center;
+  color: black;
+  background-color: transparent;
+  border: transparent;
+  margin-bottom: 25px;
+  margin-top: 25px;
+`;
+
+export const OutterBox = styled.div`
+  width: 900px;
+  margin-left: 500px;
+`;
+
+export const Backtext = styled.p`
+  padding-top: 5px;
+`;
+
+export const HeartIcon = styled(Heart)`
+  color: #333286;
+  width: 30px;
+  height: 30px;
+  fill: #333286;
+`;
+
+export const TransparentButton = styled.button`
+  background-color: transparent;
+  border: transparent;
+`;
+
+export const NavBarMovedUP = styled(NavBar)`
+  position: relative;
+`;
