+++ src/app/orderConfirmationPickUp/page.tsx
@@ -1,9 +1,7 @@
 'use client';
 
-import { ArrowLeft } from 'react-feather';
-
-import { useRouter } from 'next/navigation';
 import { useState, useEffect } from 'react';
+import BackButton from '../../components/BackButton/BackButton';
 
 import { fetchCartItems } from '../../api/supabase/queries/cart_queries';
 
@@ -11,11 +9,9 @@ import {
   FavoriteDiv,
   ColDiv,
   OutterFavoriteDiv,
-  BackDiv,
   HeaderText,
   GlobalStyle,
   OutterBox,
-  Backtext,
   Label,
   LabelBox,
   NavBarMovedUP,
@@ -29,7 +25,6 @@ import { Product } from '../../schema/schema';
 
 export default function OrderConfirmationPickUp() {
   const [Cart, setCart] = useState<Product[]>([]);
-  const router = useRouter();
 
   useEffect(() => {
     async function fetchProducts() {
@@ -43,10 +38,7 @@ export default function OrderConfirmationPickUp() {
     <div>
       <NavBarMovedUP />
       <GlobalStyle />
-      <BackDiv onClick={() => router.push('/profileScreen')}>
-        <ArrowLeft />
-        <Backtext>Back to storefront</Backtext>
-      </BackDiv>
+      <BackButton destination="./storefront" />
       <OutterBox>
         <HeaderText>Thank you, Ethan. Your order has been placed.</HeaderText>
         <OutterFavoriteDiv>
