+++ src/app/storefront/buttonValues.tsx
@@ -1,0 +1,24 @@
+export const buttons = [
+  {
+    name: 'All',
+    value: 'All',
+    count: 0,
+  },
+  {
+    name: 'Dog',
+    value: 'Dog',
+    count: 1,
+  },
+  {
+    name: 'Cat',
+    value: 'Cat',
+    count: 2,
+  },
+  {
+    name: 'Other',
+    value: 'Other',
+    count: 3,
+  },
+];
+
+export const empty = [];
