+++ src/app/storefront/productButtons.tsx
@@ -4,6 +4,8 @@ import React from 'react';
 
 import { Button, Label, IndividualContainer } from './styles';
 
+import { buttons } from './buttonValues';
+
 import {
   fetchProducts,
   filterProduct,
@@ -31,28 +33,6 @@ export default function ProductButtons(props: {
     setCategoryWord,
     index,
   } = props;
-  const buttons = [
-    {
-      name: 'All',
-      value: 'All',
-      count: 0,
-    },
-    {
-      name: 'Dog',
-      value: 'Dog',
-      count: 1,
-    },
-    {
-      name: 'Cat',
-      value: 'Cat',
-      count: 2,
-    },
-    {
-      name: 'Other',
-      value: 'Other',
-      count: 3,
-    },
-  ];
 
   async function applyFilter(
     e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
